Fatto: 1) Termina contratto -> Sistemato JDBC, ora le Entità sono indipendenti dal database

+++ Salvatore/TerminaContrattoLaptop/src/Bean/contractBean.java
@@ -11,7 +11,7 @@ package Bean;
  */
 public class contractBean {
     
-    private long contractId;
+    private int contractId;
     private int contractState;
     private String tenantNickname; 
     private String renterNickname;
@@ -21,7 +21,7 @@ public class contractBean {
     public void setTenantNickname(String tenantNickname){this.tenantNickname = tenantNickname;}
     public void setRenterNickname(String renterNickname){this.renterNickname = renterNickname;}
     
-    public long getContractId(){ return this.contractId;}
+    public int getContractId(){ return this.contractId;}
     public int getContractState(){ return this.contractState; }
     public String getTenantNickname(){ return this.tenantNickname; }
     public String getRenterNickname(){ return this.renterNickname; }
